Wrapping division in int to protect python 3

+++ src/dwell/fft/_internal.py
@@ -99,9 +99,9 @@ def rffttruncate(ft, axis):
 
     n=ft2.shape[axis]
     if odd(n):
-        x = (n-1)/2
+        x = int((n-1)/2)
     else:
-        x=n/2-1
+        x = int(n/2-1)
     d[axis]=slice(-1,x-1,-1)
     return ft2[d]
